Split MCP tool registrations into focused domain modules
Extract 1,309 lines of tool registration code from mcp.ts into 10 focused
modules under mcp/, plus a shared helpers module. mcp.ts is now a slim
session management orchestrator (~317 lines).

New modules: query-tools, tag-tools, metric-tools, activity-tools,
sync-tools, lastfm-tools, settings-tools, location-tools,
correlation-tools, trend-tools, and helpers.
