+++ apps/backend/src/mcp.ts
@@ -1,89 +1,26 @@
-/* eslint-disable max-lines -- TODO: refactor - split into mcp/ tool modules */
-import {
-  activityTypes,
-  activityTypeSchema,
-  customMetricDefinitionSchema,
-  dateOnlySchema,
-  endDateTimeQuerySchema,
-  exerciseTypeNames,
-  getExerciseTypeValue,
-  hrZoneThresholdsSchema,
-  isValidExerciseType,
-  isValidMetricOrCustom,
-  latWithValidationSchema,
-  lonWithValidationSchema,
-  startDateTimeQuerySchema,
-  syncProviderSchema,
-  validMetrics,
-  type MetricType,
-} from '@aurboda/api-spec'
+/**
+ * MCP server router with session management.
+ *
+ * Tool registrations are split into focused modules under mcp/.
+ */
 import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
 import { StreamableHTTPServerTransport } from '@modelcontextprotocol/sdk/server/streamableHttp.js'
 import { randomUUID } from 'crypto'
 import { Request, Response, Router } from 'express'
-import { z } from 'zod'
 import { Auth } from './auth'
-import {
-  deleteLastFmTagRule,
-  getAllSyncStates,
-  getLastFmTagRules,
-  getProgrammaticTags,
-  getDetectedLocations as getStoredDetectedLocations,
-  getUniqueTags,
-  getUserSettings,
-  insertLastFmTagRule,
-  upsertUserSettings,
-  type LastFmMatchMode,
-  type LastFmMatchType,
-} from './db'
-import { syncAllCalendars } from './ical-sync'
-import { syncLastFmData } from './lastfm-sync'
 import { DEFAULT_SESSION_INACTIVITY_MS, McpSessionStore } from './mcp-session-store'
+import { registerActivityTools } from './mcp/activity-tools'
+import { registerCorrelationTools } from './mcp/correlation-tools'
+import { registerLastFmTools } from './mcp/lastfm-tools'
+import { registerLocationTools } from './mcp/location-tools'
+import { registerMetricTools } from './mcp/metric-tools'
+import { registerQueryTools } from './mcp/query-tools'
+import { registerSettingsTools } from './mcp/settings-tools'
+import { registerSyncTools } from './mcp/sync-tools'
+import { registerTagTools } from './mcp/tag-tools'
+import { registerTrendTools } from './mcp/trend-tools'
 import { ouraClient } from './oura'
-import { syncAllOuraData } from './oura-sync'
-import { syncRescueTimeData } from './rescuetime-sync'
-import {
-  getActivityImpact,
-  getBaseline,
-  getEventProbability,
-  getGenericCorrelation,
-  getHrvActivitiesCorrelation,
-  type OutcomeConfig,
-  type TriggerCondition,
-} from './services/correlations'
-import { getGoalsProgress } from './services/goals'
-import {
-  deleteNamedLocation,
-  getDetectedLocations,
-  getNamedLocations,
-  insertNamedLocation,
-  updateNamedLocation,
-} from './services/locations'
-import {
-  addActivity,
-  addCustomMetric,
-  addMetric,
-  addTag,
-  deleteActivity,
-  deleteCustomMetric,
-  deleteTag,
-  getCustomMetrics,
-  updateActivity,
-} from './services/mutations'
-import {
-  BucketSize,
-  getDailySummary,
-  getPeriodSummary,
-  queryActivities,
-  queryLocations,
-  queryMetrics,
-  queryMetricsBucketed,
-  queryProductivity,
-  queryTags,
-  SyncProvider,
-} from './services/queries'
-import { getSettings, getSettingsResponse, validateAndUpdateSettings } from './services/settings'
-import { getTrend } from './services/trends'
+import { SyncProvider } from './services/queries'
 
 interface McpSession {
   transport: StreamableHTTPServerTransport
@@ -93,19 +30,6 @@ interface McpSession {
 
 type OuraClientType = ReturnType<typeof ouraClient>
 
-// ============================================================================
-// MCP-specific input schemas (reusing fields from api-spec)
-// ============================================================================
-
-// Common time range schema for MCP queries
-const mcpTimeRangeSchema = {
-  end: endDateTimeQuerySchema,
-  start: startDateTimeQuerySchema,
-}
-
-// Metric name description using validMetrics from api-spec
-const metricDescription = `Metric name. Valid metrics: ${validMetrics.join(', ')}`
-
 /**
  * Create an MCP router with optional session persistence.
  *
@@ -113,7 +37,7 @@ const metricDescription = `Metric name. Valid metrics: ${validMetrics.join(', ')
  * and can survive backend restarts. When a client reconnects with a
  * previously-issued session ID, the session is lazily restored.
  */
-// eslint-disable-next-line max-lines-per-function -- TODO: refactor - split into tool modules
+// eslint-disable-next-line max-lines-per-function -- session management logic
 export function createMcpRouter(
   auth: Auth,
   oura?: OuraClientType,
@@ -138,1228 +62,22 @@ export function createMcpRouter(
     }
   }
 
-  // Helper to create JSON text response
-  const jsonResponse = (data: unknown) => ({
-    content: [{ text: JSON.stringify(data, null, 2), type: 'text' as const }],
-  })
-
-  // Helper to create error response
-  const errorResponse = (message: string) => ({
-    content: [{ text: message, type: 'text' as const }],
-  })
-
-  // Helper to parse optional ISO date string (for fields using plain z.string())
-  // Note: Fields using startDateTimeQuerySchema/endDateTimeQuerySchema are already
-  // validated by zod, so they can be converted directly with new Date()
-  const parseOptionalDate = (dateStr: string): Date | null => {
-    const date = new Date(dateStr)
-    return isNaN(date.getTime()) ? null : date
-  }
-
-  // eslint-disable-next-line max-lines-per-function -- TODO: refactor - extract tool registration
   const createMcpServer = (user: string): McpServer => {
     const server = new McpServer({
       name: 'aurboda',
       version: '1.0.0',
     })
 
-    // ========================================================================
-    // Query Tools
-    // ========================================================================
-
-    // Tool: query_metrics
-    server.tool(
-      'query_metrics',
-      'Query health metrics for a time range. Returns time series data with timestamps and values.',
-      {
-        ...mcpTimeRangeSchema,
-        metric: z.string().describe(metricDescription),
-      },
-      async ({ end, metric, start }) => {
-        const customMetrics = await getCustomMetrics(user)
-        if (!isValidMetricOrCustom(metric, customMetrics)) {
-          const customNames = customMetrics.map((m) => m.name)
-          const allMetrics = [...validMetrics, ...customNames]
-          return errorResponse(`Invalid metric "${metric}". Valid metrics are: ${allMetrics.join(', ')}`)
-        }
-
-        // Dates are pre-validated by zod schema (startDateTimeQuerySchema/endDateTimeQuerySchema)
-        const result = await queryMetrics(user, metric, new Date(start), new Date(end), customMetrics)
-        return jsonResponse(result)
-      },
-    )
-
-    // Valid bucket sizes for bucketed metrics queries
-    const validBucketSizes = ['5m', '15m', '30m', '1h', '1d'] as const
-
-    // Tool: query_metrics_bucketed
-    server.tool(
-      'query_metrics_bucketed',
-      `Query pre-aggregated health metrics in time buckets. Returns buckets with min/max/avg/count for each metric.
-Much more efficient than query_metrics for analysis - returns ~96 buckets for a day (15m intervals) instead of 30,000+ individual samples.
-
-Use cases:
-- Daily timeline analysis: "Show me HR/HRV patterns across the day"
-- Sleep quality analysis: "How did my HRV change during sleep?"
-- Exercise response: "Compare pre/during/post exercise HR"
-- Multi-day trends: "What's my average resting HR this week?" (use 1d bucket)`,
-      {
-        ...mcpTimeRangeSchema,
-        bucket: z.enum(validBucketSizes).describe('Bucket size: "5m", "15m", "30m", "1h", or "1d"'),
-        metrics: z
-          .array(z.string())
-          .describe(`Metrics to include. Valid metrics: ${validMetrics.join(', ')}`),
-      },
-      async ({ bucket, end, metrics, start }) => {
-        const customMetrics = await getCustomMetrics(user)
-        const invalidMetrics = metrics.filter((m) => !isValidMetricOrCustom(m, customMetrics))
-        if (invalidMetrics.length > 0) {
-          const customNames = customMetrics.map((m) => m.name)
-          const allMetrics = [...validMetrics, ...customNames]
-          return errorResponse(
-            `Invalid metrics: ${invalidMetrics.join(', ')}. Valid metrics are: ${allMetrics.join(', ')}`,
-          )
-        }
-
-        // Dates are pre-validated by zod schema
-        const validatedMetrics = metrics as MetricType[]
-        const result = await queryMetricsBucketed(
-          user,
-          validatedMetrics,
-          new Date(start),
-          new Date(end),
-          bucket as BucketSize,
-        )
-        return jsonResponse(result)
-      },
-    )
-
-    // Tool: get_daily_summary
-    server.tool(
-      'get_daily_summary',
-      'Get a comprehensive summary of health data for a specific day including heart rate, steps, sleep, exercise, tags, productivity, and visited places. Also includes Oura scores (sleep_score, readiness_score, resilience_score, cardiovascular_age) when available.',
-      {
-        date: dateOnlySchema.describe('Date in YYYY-MM-DD format (e.g., 2024-01-15)'),
-      },
-      async ({ date }) => {
-        const dateMatch = date.match(/^(\d{4})-(\d{2})-(\d{2})$/)
-        if (!dateMatch) {
-          return errorResponse('Invalid date format. Use YYYY-MM-DD format.')
-        }
-
-        const dateObj = new Date(date)
-        if (isNaN(dateObj.getTime())) {
-          return errorResponse('Invalid date.')
-        }
-
-        const summary = await getDailySummary(user, dateObj, sync)
-        return jsonResponse(summary)
-      },
-    )
-
-    // Tool: query_period_summary
-    server.tool(
-      'query_period_summary',
-      'Get aggregated statistics for a time period. Returns min/max/avg/stddev for each metric, trend compared to previous period, and data completeness.',
-      {
-        ...mcpTimeRangeSchema,
-        metrics: z
-          .array(z.string())
-          .describe(`Metrics to include. Valid metrics: ${validMetrics.join(', ')}`),
-      },
-      async ({ end, metrics, start }) => {
-        const customMetrics = await getCustomMetrics(user)
-        const invalidMetrics = metrics.filter((m) => !isValidMetricOrCustom(m, customMetrics))
-        if (invalidMetrics.length > 0) {
-          const customNames = customMetrics.map((m) => m.name)
-          const allMetrics = [...validMetrics, ...customNames]
-          return errorResponse(
-            `Invalid metrics: ${invalidMetrics.join(', ')}. Valid metrics are: ${allMetrics.join(', ')}`,
-          )
-        }
-
-        // Dates are pre-validated by zod schema
-        const summary = await getPeriodSummary(user, metrics as string[], new Date(start), new Date(end))
-        return jsonResponse(summary)
-      },
-    )
-
-    // Tool: query_tags
-    server.tool(
-      'query_tags',
-      'Query tags/labels for a time range. Returns all tags with start times, optional end times, and tag text.',
-      mcpTimeRangeSchema,
-      async ({ end, start }) => {
-        // Dates are pre-validated by zod schema
-        const tags = await queryTags(user, new Date(start), new Date(end), sync)
-        return jsonResponse({ data: tags, success: true })
-      },
-    )
-
-    // Tool: query_activities
-    server.tool(
-      'query_activities',
-      'Query activities (sleep, exercise, meditation, nap) for a time range. Returns activity sessions with duration, HR zones for exercise, and other metadata.',
-      {
-        ...mcpTimeRangeSchema,
-        types: z
-          .array(activityTypeSchema)
-          .optional()
-          .describe('Activity types to include. Defaults to all types (sleep, exercise, meditation, nap).'),
-      },
-      async ({ end, start, types }) => {
-        // Dates are pre-validated by zod schema
-        // Use activityTypes from api-spec to avoid hardcoded values
-        const requestedTypes = types ?? [...activityTypes]
-        const activities = await queryActivities(user, requestedTypes, new Date(start), new Date(end), sync)
-        return jsonResponse({ data: activities, success: true })
-      },
-    )
-
-    // Tool: query_productivity
-    server.tool(
-      'query_productivity',
-      'Query productivity data (from RescueTime) for a time range. Returns application/website usage with productivity scores.',
-      mcpTimeRangeSchema,
-      async ({ end, start }) => {
-        // Dates are pre-validated by zod schema
-        const productivity = await queryProductivity(user, new Date(start), new Date(end), sync)
-        return jsonResponse({ data: productivity, success: true })
-      },
-    )
-
-    // Tool: query_locations
-    server.tool(
-      'query_locations',
-      'Query location/place visits for a time range. Returns places visited with names, coordinates, duration, and source (named, detected, or owntracks).',
-      mcpTimeRangeSchema,
-      async ({ end, start }) => {
-        // Dates are pre-validated by zod schema
-        const places = await queryLocations(user, new Date(start), new Date(end))
-        return jsonResponse({ data: places, success: true })
-      },
-    )
-
-    // ========================================================================
-    // Tag Management Tools
-    // ========================================================================
-
-    // Tool: add_tag
-    server.tool(
-      'add_tag',
-      'Add a manual tag/label to mark an activity or event. Tags can have a start time and optional end time.',
-      {
-        end_time: z
-          .string()
-          .optional()
-          .describe('Optional end time in ISO 8601 format. Omit for point-in-time tags.'),
-        merge_span: z
-          .number()
-          .int()
-          .positive()
-          .max(3600)
-          .optional()
-          .describe(
-            'If provided, merge with existing tag of same name if its end_time (or start_time for point-in-time tags) is within this many seconds of new start_time. Max 3600.',
-          ),
-        start_time: startDateTimeQuerySchema.describe(
-          'Start time in ISO 8601 format (e.g., 2024-01-15T14:30:00Z)',
-        ),
-        tag: z.string().describe('The tag/label text (e.g., "coffee", "meditation", "headache")'),
-      },
-      async ({ end_time, merge_span, start_time, tag }) => {
-        // start_time is pre-validated by zod schema
-        const startDate = new Date(start_time)
-
-        // end_time uses plain z.string() so needs manual validation
-        let endDate: Date | undefined
-        if (end_time) {
-          const parsed = parseOptionalDate(end_time)
-          if (!parsed) {
-            return errorResponse('Invalid end_time format. Use ISO 8601 format.')
-          }
-          endDate = parsed
-        }
-
-        const result = await addTag(user, {
-          endTime: endDate,
-          mergeSpan: merge_span,
-          startTime: startDate,
-          tag,
-        })
-        return jsonResponse(result)
-      },
-    )
-
-    // Tool: delete_tag
-    server.tool(
-      'delete_tag',
-      'Delete a tag by its external ID. Returns success if the tag was found and deleted.',
-      {
-        external_id: z.string().describe('The external ID of the tag to delete'),
-      },
-      async ({ external_id }) => {
-        const result = await deleteTag(user, external_id)
-        return jsonResponse(result)
-      },
-    )
-
-    // ========================================================================
-    // Metric Management Tools
-    // ========================================================================
-
-    // Tool: add_metric
-    server.tool(
-      'add_metric',
-      'Add a manual health metric measurement. Use this to log data not captured automatically.',
-      {
-        metric: z.string().describe(metricDescription),
-        time: z
-          .string()
-          .optional()
-          .describe('Measurement time in ISO 8601 format. Defaults to current time if omitted.'),
-        value: z.number().describe('The metric value (e.g., 72 for heart rate, 75.5 for weight)'),
-      },
-      async ({ metric, time, value }) => {
-        // time uses plain z.string() so needs manual validation
-        const measurementTime = time ? parseOptionalDate(time) : new Date()
-        if (!measurementTime) {
-          return errorResponse('Invalid time format. Use ISO 8601 format.')
-        }
-
-        // Validation against built-in and custom metrics is handled by addMetric service
-        const result = await addMetric(user, { metric, time: measurementTime, value })
-        if (!result.success) {
-          return errorResponse(result.error ?? 'Failed to add metric')
-        }
-        return jsonResponse(result)
-      },
-    )
-
-    // ========================================================================
-    // Custom Metric Management Tools
-    // ========================================================================
-
-    // Tool: add_custom_metric
-    server.tool(
-      'add_custom_metric',
-      'Register a new custom metric type. Custom metrics allow tracking data not covered by built-in metrics.',
-      {
-        description: z.string().optional().describe('Human-readable description of the metric'),
-        max_value: z.number().optional().describe('Maximum allowed value for validation'),
-        min_value: z.number().optional().describe('Minimum allowed value for validation'),
-        name: z
-          .string()
-          .describe(
-            'Metric name (lowercase letters, numbers, underscores). Must not conflict with built-in metrics.',
-          ),
-        unit: z.string().describe('Unit of measurement (e.g., "score", "mg", "count")'),
-      },
-      async ({ description, max_value, min_value, name, unit }) => {
-        const definition = {
-          ...(description !== undefined ? { description } : {}),
-          ...(max_value !== undefined ? { maxValue: max_value } : {}),
-          ...(min_value !== undefined ? { minValue: min_value } : {}),
-          name,
-          unit,
-        }
-
-        // Validate via the schema
-        const parsed = customMetricDefinitionSchema.safeParse(definition)
-        if (!parsed.success) {
-          return errorResponse(
-            `Invalid custom metric: ${parsed.error.issues.map((i) => i.message).join(', ')}`,
-          )
-        }
-
-        const result = await addCustomMetric(user, parsed.data)
-        if (!result.success) {
-          return errorResponse(result.error ?? 'Failed to add custom metric')
-        }
-        return jsonResponse(result)
-      },
-    )
-
-    // Tool: list_custom_metrics
-    server.tool(
-      'list_custom_metrics',
-      'List all registered custom metric types for the user.',
-      {},
-      async () => {
-        const metrics = await getCustomMetrics(user)
-        return jsonResponse({ data: metrics, success: true })
-      },
-    )
-
-    // Tool: delete_custom_metric
-    server.tool(
-      'delete_custom_metric',
-      'Delete a custom metric type. Existing data for the metric is preserved.',
-      {
-        name: z.string().describe('The name of the custom metric to delete'),
-      },
-      async ({ name }) => {
-        const result = await deleteCustomMetric(user, name)
-        return jsonResponse(result)
-      },
-    )
-
-    // ========================================================================
-    // Activity Management Tools
-    // ========================================================================
-
-    // Tool: add_activity
-    server.tool(
-      'add_activity',
-      'Add an activity session (exercise, meditation, nap). Use this to log workouts or other activities.',
-      {
-        activity_type: activityTypeSchema.describe(
-          `Type of activity. Valid types: ${activityTypes.join(', ')}`,
-        ),
-        end_time: endDateTimeQuerySchema.describe('End time in ISO 8601 format (e.g., 2024-03-15T11:45:00Z)'),
-        exercise_type: z
-          .string()
-          .optional()
-          .describe(
-            `Exercise type name (e.g., "weightlifting", "running"). Only for exercise activities. Valid types: ${exerciseTypeNames.slice(0, 10).join(', ')}...`,
-          ),
-        notes: z
-          .string()
-          .optional()
-          .describe(
-            'Activity notes. For workouts, use format: "Exercise Name: reps×weight, reps×weight" per line.',
-          ),
-        start_time: startDateTimeQuerySchema.describe(
-          'Start time in ISO 8601 format (e.g., 2024-03-15T10:30:00Z)',
-        ),
-        title: z.string().optional().describe('Activity title (e.g., "Upper body", "Morning meditation")'),
-      },
-      async ({ activity_type, end_time, exercise_type, notes, start_time, title }) => {
-        // Dates are pre-validated by zod schema
-        const startDate = new Date(start_time)
-        const endDate = new Date(end_time)
-
-        // Validate and convert exercise_type name to value
-        let data: Record<string, unknown> | undefined
-        if (exercise_type !== undefined) {
-          if (!isValidExerciseType(exercise_type)) {
-            return errorResponse(
-              `Invalid exercise_type "${exercise_type}". Valid types include: ${exerciseTypeNames.slice(0, 15).join(', ')}...`,
-            )
-          }
-          data = {
-            exerciseType: getExerciseTypeValue(exercise_type),
-            exerciseTypeName: exercise_type,
-          }
-        }
-
-        const result = await addActivity(user, {
-          activityType: activity_type,
-          data,
-          endTime: endDate,
-          notes,
-          startTime: startDate,
-          title,
-        })
-
-        if (!result.success) {
-          return errorResponse(result.error ?? 'Failed to add activity')
-        }
-
-        return jsonResponse(result)
-      },
-    )
-
-    // Tool: delete_activity
-    server.tool(
-      'delete_activity',
-      'Delete an activity by its ID. Returns success if the activity was found and deleted.',
-      {
-        id: z.string().uuid().describe('The ID of the activity to delete'),
-      },
-      async ({ id }) => {
-        const result = await deleteActivity(user, id)
-        return jsonResponse(result)
-      },
-    )
-
-    // Tool: update_activity
-    server.tool(
-      'update_activity',
-      'Update an existing activity. Can modify start_time, end_time, title, and notes. Only provided fields will be updated. Validates that end_time is after start_time (considering both new and existing values).',
-      {
-        end_time: endDateTimeQuerySchema.optional().describe('New end time in ISO 8601 format'),
-        id: z.string().uuid().describe('The ID of the activity to update'),
-        notes: z.string().optional().describe('New activity notes'),
-        start_time: startDateTimeQuerySchema.optional().describe('New start time in ISO 8601 format'),
-        title: z.string().optional().describe('New activity title'),
-      },
-      async ({ id, start_time, end_time, title, notes }) => {
-        const result = await updateActivity(user, id, {
-          endTime: end_time ? new Date(end_time) : undefined,
-          notes,
-          startTime: start_time ? new Date(start_time) : undefined,
-          title,
-        })
-
-        if (!result.success) {
-          return errorResponse(result.error ?? 'Failed to update activity')
-        }
-
-        return jsonResponse(result)
-      },
-    )
-
-    // ========================================================================
-    // Sync Tools
-    // ========================================================================
-
-    // Tool: sync_oura
-    server.tool(
-      'sync_oura',
-      'Sync data from Oura Ring API. Fetches cardiovascular age, readiness, resilience, sleep scores, meditation sessions, and tags.',
-      {
-        full_resync: z
-          .boolean()
-          .optional()
-          .describe(
-            'If true, fetches all historical data (default 90 days). Otherwise, fetches only since last sync.',
-          ),
-        start_date: z
-          .string()
-          .optional()
-          .describe('Optional start date for sync in YYYY-MM-DD format. Only used with full_resync.'),
-      },
-      async ({ full_resync, start_date }) => {
-        if (!oura) {
-          return errorResponse('Oura integration is not configured on this server.')
-        }
-
-        try {
-          const results = await syncAllOuraData(user, oura, {
-            fullResync: full_resync,
-            startDate: start_date ? new Date(start_date) : undefined,
-          })
-
-          const summary = results.map((r) => ({
-            dataType: r.dataType,
-            error: r.error,
-            recordsProcessed: r.recordsProcessed,
-            status: r.status,
-          }))
-
-          return jsonResponse({ results: summary, success: true })
-        } catch (error) {
-          const message = error instanceof Error ? error.message : 'Unknown error'
-          return jsonResponse({ error: message, success: false })
-        }
-      },
-    )
-
-    // Tool: sync_rescuetime
-    server.tool(
-      'sync_rescuetime',
-      'Sync productivity data from RescueTime API. Fetches application and website usage with productivity scores.',
-      {
-        full_resync: z
-          .boolean()
-          .optional()
-          .describe(
-            'If true, fetches all historical data (default 30 days). Otherwise, fetches only since last sync.',
-          ),
-        start_date: z
-          .string()
-          .optional()
-          .describe('Optional start date for sync in YYYY-MM-DD format. Only used with full_resync.'),
-      },
-      async ({ full_resync, start_date }) => {
-        const settings = await getSettings(user)
-        if (!settings.rescueTimeKey) {
-          return errorResponse('RescueTime API key is not configured in user settings.')
-        }
-
-        try {
-          const result = await syncRescueTimeData(user, settings.rescueTimeKey, {
-            fullResync: full_resync,
-            startDate: start_date ? new Date(start_date) : undefined,
-          })
-
-          return jsonResponse({
-            error: result.error,
-            recordsProcessed: result.recordsProcessed,
-            status: result.status,
-            success: result.status === 'success',
-          })
-        } catch (error) {
-          const message = error instanceof Error ? error.message : 'Unknown error'
-          return jsonResponse({ error: message, success: false })
-        }
-      },
-    )
-
-    // Tool: sync_calendars
-    server.tool(
-      'sync_calendars',
-      'Sync events from configured calendar ICS URLs. Fetches ICS data and stores events as tags for correlation analysis.',
-      {
-        full_resync: z
-          .boolean()
-          .optional()
-          .describe('If true, re-fetches all events. Otherwise, performs a normal sync.'),
-      },
-      async () => {
-        const settings = await getSettings(user)
-        if (!settings.calendars || settings.calendars.length === 0) {
-          return errorResponse('No calendars configured in user settings. Add calendars in Settings first.')
-        }
-
-        try {
-          const results = await syncAllCalendars(user, settings.calendars)
-
-          const summary = results.map((r) => ({
-            calendar: r.calendar,
-            error: r.error,
-            eventsProcessed: r.eventsProcessed,
-            status: r.status,
-          }))
-
-          return jsonResponse({ results: summary, success: true })
-        } catch (error) {
-          const message = error instanceof Error ? error.message : 'Unknown error'
-          return jsonResponse({ error: message, success: false })
-        }
-      },
-    )
-
-    // Tool: sync_lastfm
-    server.tool(
-      'sync_lastfm',
-      'Sync scrobbles from Last.fm. Fetches recent tracks and applies auto-tagging rules.',
-      {
-        full_resync: z
-          .boolean()
-          .optional()
-          .describe(
-            'If true, fetches all historical data (default 30 days). Otherwise, fetches only since last sync.',
-          ),
-        start_date: z
-          .string()
-          .optional()
-          .describe('Optional start date for sync in YYYY-MM-DD format. Only used with full_resync.'),
-      },
-      async ({ full_resync, start_date }) => {
-        const lastFmApiKey = process.env.LASTFM_API_KEY
-        if (!lastFmApiKey) {
-          return errorResponse('Last.fm API key is not configured on this server.')
-        }
-
-        const settings = await getSettings(user)
-        if (!settings.lastFmUsername) {
-          return errorResponse('Last.fm username is not configured in user settings.')
-        }
-
-        try {
-          const result = await syncLastFmData(user, lastFmApiKey, settings.lastFmUsername, {
-            fullResync: full_resync,
-            startDate: start_date ? new Date(start_date) : undefined,
-          })
-
-          return jsonResponse({
-            error: result.error,
-            scrobblesProcessed: result.scrobblesProcessed,
-            status: result.status,
-            success: result.status === 'success',
-            tagsCreated: result.tagsCreated,
-          })
-        } catch (error) {
-          const message = error instanceof Error ? error.message : 'Unknown error'
-          return jsonResponse({ error: message, success: false })
-        }
-      },
-    )
-
-    // Tool: get_sync_status
-    server.tool(
-      'get_sync_status',
-      'Get the current sync status for Oura, RescueTime, Calendar, and Last.fm data sources. Shows last sync time, status, and any errors.',
-      {
-        provider: syncProviderSchema.optional().describe('Which provider to check. Defaults to "all".'),
-      },
-      async ({ provider = 'all' }) => {
-        try {
-          const states: Record<string, unknown[]> = {}
-
-          if (provider === 'all' || provider === 'oura') {
-            states.oura = await getAllSyncStates(user, 'oura')
-          }
-
-          if (provider === 'all' || provider === 'rescuetime') {
-            states.rescuetime = await getAllSyncStates(user, 'rescuetime')
-          }
-
-          if (provider === 'all' || provider === 'calendar') {
-            states.calendar = await getAllSyncStates(user, 'calendar')
-          }
-
-          if (provider === 'all' || provider === 'lastfm') {
-            states.lastfm = await getAllSyncStates(user, 'lastfm')
-          }
-
-          return jsonResponse({ states, success: true })
-        } catch (error) {
-          const message = error instanceof Error ? error.message : 'Unknown error'
-          return jsonResponse({ error: message, success: false })
-        }
-      },
-    )
-
-    // ========================================================================
-    // Last.fm Tag Rules Tools
-    // ========================================================================
-
-    // Tool: get_lastfm_tag_rules
-    server.tool(
-      'get_lastfm_tag_rules',
-      'Get all Last.fm auto-tagging rules. These rules create tags when scrobbles match specified criteria.',
-      {},
-      async () => {
-        const rules = await getLastFmTagRules(user)
-        const serialized = rules.map((r) => ({
-          ...r,
-          createdAt: r.createdAt.toISOString(),
-        }))
-        return jsonResponse({ data: serialized, success: true })
-      },
-    )
-
-    // Tool: add_lastfm_tag_rule
-    server.tool(
-      'add_lastfm_tag_rule',
-      'Add a Last.fm auto-tagging rule. Creates tags when scrobbles match the specified criteria.',
-      {
-        artist_name: z
-          .string()
-          .optional()
-          .describe('Artist name to match (required for artist or track_artist match type)'),
-        match_mode: z
-          .enum(['exact', 'contains'])
-          .optional()
-          .describe('Match mode: exact (case-insensitive) or contains (substring). Default: exact'),
-        match_type: z
-          .enum(['track', 'artist', 'track_artist'])
-          .describe(
-            'Type of match: track (any track with name), artist (any track by artist), track_artist (exact track + artist)',
-          ),
-        rule_name: z.string().min(1).describe('Human-readable name for the rule'),
-        tag_name: z.string().min(1).describe('Tag to create when rule matches'),
-        track_name: z
-          .string()
-          .optional()
-          .describe('Track name to match (required for track or track_artist match type)'),
-      },
-      async ({ artist_name, match_mode, match_type, rule_name, tag_name, track_name }) => {
-        // Validate required fields based on match_type
-        if ((match_type === 'track' || match_type === 'track_artist') && !track_name) {
-          return errorResponse(`track_name is required for match_type "${match_type}"`)
-        }
-        if ((match_type === 'artist' || match_type === 'track_artist') && !artist_name) {
-          return errorResponse(`artist_name is required for match_type "${match_type}"`)
-        }
-
-        try {
-          const rule = await insertLastFmTagRule(user, {
-            artistName: artist_name,
-            matchMode: (match_mode ?? 'exact') as LastFmMatchMode,
-            matchType: match_type as LastFmMatchType,
-            ruleName: rule_name,
-            tagName: tag_name,
-            trackName: track_name,
-          })
-
-          return jsonResponse({
-            data: {
-              ...rule,
-              createdAt: rule.createdAt.toISOString(),
-            },
-            success: true,
-          })
-        } catch (error) {
-          const message = error instanceof Error ? error.message : 'Unknown error'
-          // Check for unique constraint violation
-          if (message.includes('unique_rule')) {
-            return errorResponse('A rule with the same match criteria and tag already exists')
-          }
-          return jsonResponse({ error: message, success: false })
-        }
-      },
-    )
-
-    // Tool: delete_lastfm_tag_rule
-    server.tool(
-      'delete_lastfm_tag_rule',
-      'Delete a Last.fm auto-tagging rule by its ID.',
-      {
-        rule_id: z.string().uuid().describe('The ID of the rule to delete'),
-      },
-      async ({ rule_id }) => {
-        const deleted = await deleteLastFmTagRule(user, rule_id)
-        if (!deleted) {
-          return jsonResponse({ error: 'Rule not found', success: false })
-        }
-        return jsonResponse({ success: true })
-      },
-    )
-
-    // ========================================================================
-    // User Settings Tools
-    // ========================================================================
-
-    // Tool: get_user_settings
-    server.tool(
-      'get_user_settings',
-      'Get user settings including birth date and effective HR zones. HR zones are used to calculate time spent in different heart rate zones during exercise.',
-      {},
-      async () => {
-        const result = await getSettingsResponse(user)
-        return jsonResponse(result)
-      },
-    )
-
-    // Tool: update_user_settings
-    server.tool(
-      'update_user_settings',
-      'Update user settings. Can set birth date (for age-based HR zones) and/or custom HR zone thresholds.',
-      {
-        birth_date: z
-          .string()
-          .nullable()
-          .optional()
-          .describe('Birth date in YYYY-MM-DD format. Set to null to clear.'),
-        hr_zone_start: hrZoneThresholdsSchema
-          .nullable()
-          .optional()
-          .describe('Custom HR zone start thresholds. Values must be ascending. Set to null to clear.'),
-      },
-      async ({ birth_date, hr_zone_start }) => {
-        const result = await validateAndUpdateSettings(user, {
-          birthDate: birth_date,
-          hrZoneStart: hr_zone_start,
-        })
-        return jsonResponse(result)
-      },
-    )
-
-    // Tool: get_unique_tags
-    server.tool(
-      'get_unique_tags',
-      'Get all unique tag names that have been recorded. Returns a list of tag strings.',
-      {},
-      async () => {
-        const tags = await getUniqueTags(user)
-        return jsonResponse({ data: tags, success: true })
-      },
-    )
-
-    // Tool: get_programmatic_tags
-    server.tool(
-      'get_programmatic_tags',
-      'Get all programmatic tags (UUIDs, tag_* prefixes) with their current mapped names. These are tags that look like they need human-readable display names. Tags without a currentName are unmapped.',
-      {},
-      async () => {
-        const tags = await getProgrammaticTags(user)
-        const settings = await getUserSettings(user)
-        const mappings = settings?.tagMappings ?? {}
-
-        const data = tags.map((tag) => ({
-          count: tag.count,
-          currentName: mappings[tag.tagKey] ?? null,
-          latestTime: tag.latestTime.toISOString(),
-          tagKey: tag.tagKey,
-        }))
-
-        return jsonResponse({ data, success: true })
-      },
-    )
-
-    // Tool: set_tag_mapping
-    server.tool(
-      'set_tag_mapping',
-      'Set a display name for a programmatic tag (UUID, tag_* prefix, etc.). Use after get_programmatic_tags to name unmapped tags.',
-      {
-        name: z.string().min(1).describe('Display name for the tag'),
-        tag_key: z.string().min(1).describe('The programmatic tag identifier (UUID, tag_* prefix, etc.)'),
-      },
-      async ({ name, tag_key }) => {
-        const settings = await getUserSettings(user)
-        const currentMappings = settings?.tagMappings ?? {}
-        const newMappings = { ...currentMappings, [tag_key]: name }
-
-        await upsertUserSettings(user, { tagMappings: newMappings })
-
-        return jsonResponse({ mapping: newMappings, success: true })
-      },
-    )
-
-    // Tool: get_tag_mappings
-    server.tool(
-      'get_tag_mappings',
-      'Get all current tag mappings (tag key -> display name).',
-      {},
-      async () => {
-        const settings = await getUserSettings(user)
-        return jsonResponse({ mappings: settings?.tagMappings ?? {}, success: true })
-      },
-    )
-
-    // Tool: get_goal_progress
-    server.tool(
-      'get_goal_progress',
-      'Get progress toward all user goals. Returns current value, min/max targets, and how much will be lost when the oldest day exits the rolling window.',
-      {},
-      async () => {
-        const goals = await getGoalsProgress(user)
-        return jsonResponse({ goals, success: true })
-      },
-    )
-
-    // ========================================================================
-    // Location Tools
-    // ========================================================================
-
-    // Tool: get_named_locations
-    server.tool(
-      'get_named_locations',
-      'List all named locations. These are user-defined places with names and coordinates.',
-      {},
-      async () => {
-        const locations = await getNamedLocations(user)
-        return jsonResponse({ data: locations, success: true })
-      },
-    )
-
-    // Tool: get_detected_locations
-    server.tool(
-      'get_detected_locations',
-      'Get frequently visited locations that are not yet named. Detects places where user spent 60+ minutes. Returns coordinates, visit count, and total time spent.',
-      {
-        ...mcpTimeRangeSchema,
-        min_duration: z.number().optional().describe('Minimum stay duration in minutes. Defaults to 60.'),
-      },
-      async ({ end, min_duration, start }) => {
-        // Dates are pre-validated by zod schema
-        const detected = await getDetectedLocations(user, {
-          end: new Date(end),
-          minDurationMinutes: min_duration,
-          start: new Date(start),
-        })
-
-        return jsonResponse({ data: detected, success: true })
-      },
-    )
-
-    // Tool: get_stored_detected_locations
-    server.tool(
-      'get_stored_detected_locations',
-      'Get stored detected locations with geocoded addresses. These are locations that have been previously detected and stored in the database.',
-      {},
-      async () => {
-        const detected = await getStoredDetectedLocations(user)
-        // JSON.stringify automatically converts Date objects to ISO strings
-        return jsonResponse({ data: detected, success: true })
-      },
-    )
-
-    // Tool: add_named_location
-    server.tool(
-      'add_named_location',
-      'Create a named location. Use this to save a frequently visited place with a name.',
-      {
-        lat: latWithValidationSchema.describe('Latitude of the location (-90 to 90)'),
-        lon: lonWithValidationSchema.describe('Longitude of the location (-180 to 180)'),
-        name: z.string().describe('Name for the location (e.g., "Home", "Office", "Gym")'),
-        radius: z.number().optional().describe('Radius in meters. Defaults to 200.'),
-      },
-      async ({ lat, lon, name, radius }) => {
-        // lat/lon are pre-validated by zod schema (latWithValidationSchema/lonWithValidationSchema)
-        const location = await insertNamedLocation(user, { lat, lon, name, radius })
-        return jsonResponse({ data: location, success: true })
-      },
-    )
-
-    // Tool: update_named_location
-    server.tool(
-      'update_named_location',
-      'Update an existing named location. Can change name, coordinates, or radius.',
-      {
-        id: z.string().describe('The ID of the named location to update'),
-        lat: z.number().optional().describe('New latitude (-90 to 90). Must be provided with lon.'),
-        lon: z.number().optional().describe('New longitude (-180 to 180). Must be provided with lat.'),
-        name: z.string().optional().describe('New name for the location'),
-        radius: z.number().optional().describe('New radius in meters'),
-      },
-      async ({ id, lat, lon, name, radius }) => {
-        if ((lat !== undefined && lon === undefined) || (lon !== undefined && lat === undefined)) {
-          return errorResponse('lat and lon must be provided together.')
-        }
-
-        if (lat !== undefined && (lat < -90 || lat > 90)) {
-          return errorResponse('Invalid latitude. Must be between -90 and 90.')
-        }
-        if (lon !== undefined && (lon < -180 || lon > 180)) {
-          return errorResponse('Invalid longitude. Must be between -180 and 180.')
-        }
-
-        const location = await updateNamedLocation(user, id, { lat, lon, name, radius })
-        if (!location) {
-          return jsonResponse({ error: 'Named location not found', success: false })
-        }
-        return jsonResponse({ data: location, success: true })
-      },
-    )
-
-    // Tool: delete_named_location
-    server.tool(
-      'delete_named_location',
-      'Delete a named location by its ID.',
-      {
-        id: z.string().describe('The ID of the named location to delete'),
-      },
-      async ({ id }) => {
-        const deleted = await deleteNamedLocation(user, id)
-        if (!deleted) {
-          return jsonResponse({ error: 'Named location not found', success: false })
-        }
-        return jsonResponse({ success: true })
-      },
-    )
-
-    // Tool: promote_detected_location
-    server.tool(
-      'promote_detected_location',
-      'Create a named location from detected coordinates. Use after get_detected_locations to save a frequently visited place.',
-      {
-        lat: latWithValidationSchema.describe('Latitude from detected location'),
-        lon: lonWithValidationSchema.describe('Longitude from detected location'),
-        name: z.string().describe('Name for the location'),
-        radius: z.number().optional().describe('Radius in meters. Uses suggested radius if not provided.'),
-      },
-      async ({ lat, lon, name, radius }) => {
-        const location = await insertNamedLocation(user, { lat, lon, name, radius })
-        return jsonResponse({ data: location, success: true })
-      },
-    )
-
-    // ========================================================================
-    // Correlation Analysis Tools
-    // ========================================================================
-
-    // Tool: get_baseline
-    server.tool(
-      'get_baseline',
-      'Get HRV baseline statistics (7-day and 30-day averages). Returns mean HRV (rmssd) and resting heart rate with trend percentage.',
-      {
-        reference_date: dateOnlySchema
-          .optional()
-          .describe('Reference date for baseline calculation in YYYY-MM-DD format. Defaults to today.'),
-      },
-      async ({ reference_date }) => {
-        const referenceDate = reference_date ? new Date(reference_date) : undefined
-        const baseline = await getBaseline(user, referenceDate)
-        return jsonResponse({ data: baseline, success: true })
-      },
-    )
-
-    // Tool: get_hrv_activities_correlation
-    server.tool(
-      'get_hrv_activities_correlation',
-      'Get HRV correlations with various activities. Returns Pearson correlation coefficients between HRV and productivity, locations, activities, and tags.',
-      {
-        period_days: z.number().int().optional().describe('Number of days to analyze. Defaults to 30.'),
-      },
-      async ({ period_days }) => {
-        const periodDays = period_days ?? 30
-        const correlations = await getHrvActivitiesCorrelation(user, periodDays, sync)
-        return jsonResponse({ data: correlations, success: true })
-      },
-    )
-
-    // Tool: get_activity_impact
-    server.tool(
-      'get_activity_impact',
-      'Get the impact of a specific activity/tag on HRV and heart rate. Compares metric values before, during, and after the activity using time windows.',
-      {
-        activity: z
-          .string()
-          .describe('The activity or tag name to analyze (e.g., "gym", "coffee", "meditation")'),
-        activity_type: z
-          .enum(['productivity_category', 'productivity_app', 'location', 'tag', 'activity_type'])
-          .describe('Type of activity to search for'),
-        period_days: z.number().int().optional().describe('Number of days to analyze. Defaults to 90.'),
-        window_minutes: z
-          .number()
-          .int()
-          .optional()
-          .describe('Minutes to analyze before/after the activity. Defaults to 30.'),
-      },
-      async ({ activity, activity_type, period_days, window_minutes }) => {
-        const periodDays = period_days ?? 90
-        const windowMinutes = window_minutes ?? 30
-
-        const impact = await getActivityImpact(user, activity, activity_type, windowMinutes, periodDays, sync)
-        return jsonResponse({ data: impact, success: true })
-      },
-    )
-
-    // Tool: get_event_probability
-    server.tool(
-      'get_event_probability',
-      'Get the probability correlation between two events. Analyzes whether one event (trigger) increases or decreases the probability of another event (outcome) occurring within specified time windows. Uses chi-squared test for statistical significance.',
-      {
-        lag_windows: z
-          .array(z.string())
-          .optional()
-          .describe(
-            'Time windows to analyze (e.g., ["12h", "24h", "36h", "48h"]). Uses hours (h) or days (d).',
-          ),
-        outcome_pattern: z
-          .string()
-          .describe('Regex pattern for outcome tags (e.g., "headache|migraine", "good_sleep")'),
-        period_days: z.number().int().optional().describe('Number of days to analyze. Defaults to 365.'),
-        trigger_type: z.enum(['activity', 'tag']).describe('Type of trigger event'),
-        trigger_value: z
-          .string()
-          .describe('Trigger activity type or tag pattern (e.g., "exercise", "gym", "coffee")'),
-      },
-      async ({ lag_windows, outcome_pattern, period_days, trigger_type, trigger_value }) => {
-        const probability = await getEventProbability(
-          user,
-          { type: trigger_type, value: trigger_value },
-          { pattern: outcome_pattern, type: 'tag' },
-          lag_windows ?? ['12h', '24h', '36h', '48h'],
-          period_days ?? 365,
-          sync,
-        )
-        return jsonResponse({ data: probability, success: true })
-      },
-    )
-
-    // Tool: get_generic_correlation
-    server.tool(
-      'get_generic_correlation',
-      `Analyze correlations between compound triggers and various outcomes. Supports:
-- Multiple trigger conditions with AND logic (e.g., "exercise 3x AND fatcoffee 5x in a week")
-- Different outcome types: tags, metrics (weight, body_fat, etc.), or productivity time
-- Rolling windows for trigger counting
-Examples:
-- "Does meditation correlate with more productive time?" -> trigger: tag "meditation", outcome: productivity
-- "When I exercise 3x and tag FatCoffee 5x in a week, does my weight change?" -> compound triggers, metric outcome`,
-      {
-        lag_windows: z
-          .array(z.string())
-          .optional()
-          .describe(
-            'Time windows to analyze after trigger (e.g., ["24h", "7d"]). Uses hours (h) or days (d).',
-          ),
-        outcome: z
-          .object({
-            app: z.string().optional().describe('For productivity: specific app to measure'),
-            category: z.string().optional().describe('For productivity: category to measure'),
-            metric: z.string().optional().describe('For metric: metric name (e.g., "weight", "body_fat")'),
-            pattern: z.string().optional().describe('For tag: regex pattern to match'),
-            type: z.enum(['tag', 'metric', 'productivity']).describe('Type of outcome to measure'),
-          })
-          .describe('Outcome to measure'),
-        period_days: z.number().int().optional().describe('Number of days to analyze. Defaults to 90.'),
-        triggers: z
-          .array(
-            z.object({
-              minCount: z.number().int().optional().describe('Minimum occurrences in window (default: 1)'),
-              pattern: z.string().describe('Pattern to match (regex for tags, name for activities)'),
-              type: z
-                .enum(['activity', 'tag', 'productivity_category', 'productivity_app'])
-                .describe('Type of trigger'),
-              windowDays: z.number().int().optional().describe('Rolling window in days (default: 1)'),
-            }),
-          )
-          .describe('Trigger conditions (all must be met)'),
-      },
-      async ({ lag_windows, outcome, period_days, triggers }) => {
-        const result = await getGenericCorrelation(
-          user,
-          triggers as TriggerCondition[],
-          outcome as OutcomeConfig,
-          lag_windows ?? ['24h', '48h', '7d'],
-          period_days ?? 90,
-          sync,
-        )
-        return jsonResponse({ data: result, success: true })
-      },
-    )
-
-    // ========================================================================
-    // Trend Analysis Tools
-    // ========================================================================
-
-    // Tool: get_trend
-    server.tool(
-      'get_trend',
-      `Calculate time-weighted trend for tags or metrics using Exponential Moving Average (EMA).
-EMA gives more weight to recent data while smoothing over a configurable period.
-The half-life parameter controls decay: after half-life days, an event's weight drops to 50%.
-
-Examples:
-- Tag trend: "How many painkillers per month?" -> pattern: "pain_killer", sourceType: "tag"
-- Metric trend: "What's my average weight trend?" -> pattern: "weight", sourceType: "metric"
-
-Common half-life values:
-- 7 days (quick): Responds to changes within a week
-- 15 days (responsive): Balanced, good default
-- 30 days (stable): Smooths out short-term variation`,
-      {
-        aggregation: z
-          .enum(['count', 'sum', 'mean'])
-          .optional()
-          .describe('Aggregation method. For tags: "count" (default). For metrics: "mean" or "sum".'),
-        display_period: z
-          .enum(['daily', 'weekly', 'monthly'])
-          .optional()
-          .describe('Period to normalize rate to. Default: "monthly".'),
-        half_life_days: z
-          .number()
-          .positive()
-          .optional()
-          .describe(
-            'EMA half-life in days. Default: 15. Common values: 7 (quick), 15 (responsive), 30 (stable).',
-          ),
-        lookback_days: z
-          .number()
-          .positive()
-          .optional()
-          .describe('Days of historical data to include. Default: 90.'),
-        pattern: z
-          .string()
-          .describe(
-            'For tags: regex pattern to match (e.g., "pain_killer|headache"). For metrics: metric name.',
-          ),
-        source_type: z.enum(['tag', 'metric']).describe('Type of data source: "tag" or "metric".'),
-      },
-      async ({ aggregation, display_period, half_life_days, lookback_days, pattern, source_type }) => {
-        try {
-          const result = await getTrend(user, {
-            aggregation,
-            displayPeriod: display_period,
-            halfLifeDays: half_life_days,
-            lookbackDays: lookback_days,
-            pattern,
-            sourceType: source_type,
-          })
-          return jsonResponse({ data: result, success: true })
-        } catch (error) {
-          const message = error instanceof Error ? error.message : 'Unknown error'
-          return jsonResponse({ error: message, success: false })
-        }
-      },
-    )
+    registerQueryTools(server, user, sync)
+    registerTagTools(server, user)
+    registerMetricTools(server, user)
+    registerActivityTools(server, user)
+    registerSyncTools(server, user, oura)
+    registerLastFmTools(server, user)
+    registerSettingsTools(server, user)
+    registerLocationTools(server, user)
+    registerCorrelationTools(server, user, sync)
+    registerTrendTools(server, user)
 
     return server
   }
